Refactored bottom sheets for tests
